requirements.txt
# SVAMS Requirements
# Install: pip install -r requirements.txt

# Flask Backend
flask==3.0.3
flask-jwt-extended==4.6.0
mysql-connector-python==8.4.0
bcrypt==4.1.3
python-dotenv==1.0.1

# Vision — YOLO + OCR
ultralytics>=8.2.0        # YOLOv8
easyocr>=1.7.1            # EasyOCR
opencv-python>=4.9.0      # cv2 (use opencv-python-headless on servers)
numpy>=1.26.0

# Streamlit Frontend
streamlit==1.36.0
requests==2.32.3
pandas==2.2.2
plotly==5.22.0
